Create useCloudImages hook
